refactored SortedMap factors organization
git-svn-id: file:///home/SUBVERSION/jas/trunk@2285 eafcb37d-f50b-0410-b8f7-d4cb9effa109

+++ src/edu/jas/ufd/FactorAlgebraic.java
@@ -189,7 +189,7 @@ public class FactorAlgebraic <C extends GcdRingElem<C>>
             k++;
         }
         // Res is now squarefree, so we can factor it
-        SortedMap<GenPolynomial<C>,Integer> nfacs = factorCoeff.baseFactors( res );
+        SortedMap<GenPolynomial<C>,Long> nfacs = factorCoeff.baseFactors( res );
         System.out.println("\nnfacs = " + nfacs); // Q[X]
         if ( !factorCoeff.isFactorization( res, nfacs ) ) {
            System.out.println("isFactorization = false");
